+++ apps/docs/src/.vitepress/config.ts
@@ -1,17 +1,33 @@
+import { fileURLToPath } from "node:url";
 import { defineConfig } from "vitepress";
 import packageJson from "../../../../packages/sit-onyx/package.json";
 import { getComponents } from "./utils";
 
+const componentNames = await getComponents();
+
 // https://vitepress.dev/reference/site-config
 export default defineConfig({
+  vite: {
+    resolve: {
+      alias: {
+        "~components": getFilePath("../../../../packages/sit-onyx/src/components"),
+      },
+    },
+  },
   title: "onyx",
-  description: "Vue.js component library and design system",
+  description: packageJson.description,
+  head: [
+    ["link", { rel: "icon", href: "/favicon.svg" }],
+    ["link", { rel: "apple-touch-icon", href: "/apple-touch-icon.png", sizes: "180x180" }],
+    ["link", { rel: "manifest", href: "/site.webmanifest" }],
+    ["link", { rel: "mask-icon", href: "/safari-pinned-tab.svg", color: "#0c1f2d" }],
+    ["meta", { name: "msapplication-TileColor", content: "#0c1f2d" }],
+    ["meta", { name: "theme-color", content: "#0c1f2d" }],
+  ],
   themeConfig: {
     externalLinkIcon: true,
-    logo: {
-      light: "/logo-light.svg",
-      dark: "/logo-dark.svg",
-    },
+    logo: "/logo.svg",
+    siteTitle: false,
     footer: {
       message: "Released under the Apache-2.0 License.",
       copyright: "Copyright © 2023-present Schwarz IT KG",
@@ -24,11 +40,15 @@ export default defineConfig({
     },
     lastUpdated: {}, // needed to show the last updated text with default settings
     nav: [
-      { text: "Brand", link: "/brand/", activeMatch: "/brand/" },
+      { text: "The Team", link: "/brand/team", activeMatch: "/brand/" },
       { text: "Basics", link: "/basics/", activeMatch: "/basics/" },
-      { text: "Tokens", link: "/tokens/", activeMatch: "/tokens/" },
+      { text: "Tokens", link: "/tokens/introduction", activeMatch: "/tokens/" },
       { text: "Development", link: "/development/", activeMatch: "/development/" },
-      { text: "Resources", link: "/resources/", activeMatch: "/resources/" },
+      {
+        text: "Resources",
+        activeMatch: "/resources/",
+        items: [{ text: "Icons", link: "/resources/icons" }],
+      },
       { text: "Report a bug", link: packageJson.bugs.url },
       { text: "Q&A", link: "https://github.com/schwarzit/onyx/discussions/categories/q-a" },
     ],
@@ -36,14 +56,8 @@ export default defineConfig({
     sidebar: {
       "/brand": [
         {
-          items: [
-            { text: "The Team", link: "/brand/team" },
-            { text: "Ideology", link: "/brand/ideology" },
-            { text: "Principles", link: "/brand/principles" },
-            { text: "Dependencies", link: "/brand/dependencies" },
-            { text: "Roadmap", link: "/brand/roadmap" },
-            { text: "Changelog", link: "/brand/changelog" },
-          ],
+          text: "Brand",
+          items: [{ text: "The Team", link: "/brand/team" }],
         },
       ],
       "/basics": [
@@ -57,7 +71,7 @@ export default defineConfig({
             { text: "Layout", link: "/layout" },
             { text: "Units", link: "/units" },
             { text: "Motion", link: "/motion" },
-            { text: "Component states", link: "/states" },
+            { text: "States", link: "/states" },
             { text: "Density", link: "/density" },
             { text: "Truncation", link: "/truncation" },
             { text: "Elevation", link: "/elevation" },
@@ -82,6 +96,7 @@ export default defineConfig({
       ],
       "/tokens": [
         {
+          text: "Design Tokens",
           base: "/tokens",
           items: [
             { text: "Introduction", link: "/introduction" },
@@ -101,13 +116,14 @@ export default defineConfig({
             { text: "i18n", link: "/i18n" },
             { text: "Grid", link: "/grid" },
             { text: "Typography", link: "/typography" },
+            { text: "Changelog", link: "/packages/changelogs/sit-onyx" },
           ],
         },
         {
           text: "Components",
           base: "/development",
           collapsed: false,
-          items: getComponents().map((name) => ({ text: name, link: `/${name}` })),
+          items: componentNames.map((name) => ({ text: name, link: `/${name}` })),
         },
         {
           text: "Other onyx npm packages",
@@ -116,6 +132,7 @@ export default defineConfig({
           items: [
             { text: "Figma utilities", link: "/figma-utils" },
             { text: "Headless composables", link: "/headless" },
+            { text: "Icons", link: "/icons" },
             { text: "Storybook utilities", link: "/storybook-utils" },
             { text: "VitePress theme", link: "/vitepress-theme" },
           ],
@@ -123,9 +140,16 @@ export default defineConfig({
       ],
       "/resources": [
         {
-          items: [{ text: "test", link: "/resources" }],
+          text: "Resources",
+          base: "/resources",
+          items: [{ text: "Icons", link: "/icons" }],
         },
       ],
     },
   },
 });
+
+/** Gets the given path while ensuring cross-platform and correct decoding */
+function getFilePath(path: string) {
+  return fileURLToPath(new URL(path, import.meta.url));
+}